sso_cookies.py: Parse according to RFC 6265
Individual cookies in the `Cookie` header, according to Section 5.4 of RFC 6265 [1],
are separated by the "semicolon-space" (`; `) combination. This becomes relevant
when there is a cookie whose value contains a semicolon. I have just encountered one
on O'Reilly Learning, namely `OptanonConsent` ending in `&geolocation=US;CA`,
which was breaking the script.

[1]: https://www.rfc-editor.org/rfc/rfc6265#section-5.4